CR: rename files to kebab-case
