+++ src/index.js
@@ -6,5 +6,5 @@ export {queries}
 
 export * from './queries'
 export * from './wait'
-export * from './waitForElement'
+export * from './wait-for-element'
 export * from './matches'
